+++ src/mailman/mta/tests/test_postfix.py
@@ -17,6 +17,7 @@
 
 """Test Postfix config."""
 
+import os
 import unittest
 
 from mailman.config import config
@@ -27,11 +28,53 @@ from mailman.testing.layers import ConfigLayer
 class TestPostfixConfig(unittest.TestCase):
     layer = ConfigLayer
 
-    def test_regex_config(self):
-        config.push('regex_config', """\
+    def test_type_default_config(self):
+        config.push('database_type_config', """\
 [mta]
-configuration: python:mailman.mta.tests.data.postfix
+configuration: python:mailman.mta.tests.data.postfix_type_default
 """)
         lmtp = LMTP()
-        self.assertEqual('regex', lmtp.transport_file_type)
-        config.pop('regex_config')
+        lmtp.regenerate()
+        for db_basename in ('postfix_lmtp', 'postfix_domains'):
+            src_path = f'{config.DATA_DIR}/{db_basename}'
+            db_path = f'{src_path}.created-db'
+            self.assertTrue(
+                os.path.isfile(db_path),
+                f'Created database file: {db_path}'
+            )
+            with open(db_path, 'r') as f:
+                self.assertEqual(
+                    src_path, f.read().rstrip('\n'),
+                    'Command-line argument(s) for postmap command'
+                )
+        config.pop('database_type_config')
+
+    def test_type_hash_config(self):
+        config.push('database_type_config', """\
+[mta]
+configuration: python:mailman.mta.tests.data.postfix_type_hash
+""")
+        lmtp = LMTP()
+        lmtp.regenerate()
+        for db_basename in ('postfix_lmtp', 'postfix_domains'):
+            src_path = f'{config.DATA_DIR}/{db_basename}'
+            db_path = f'{src_path}.created-db'
+            self.assertTrue(
+                os.path.isfile(db_path),
+                f'Created database file: {db_path}'
+            )
+            with open(db_path, 'r') as f:
+                self.assertEqual(
+                    f'hash:{src_path}', f.read().rstrip("\n"),
+                    'Command-line argument(s) for postmap command'
+                )
+        config.pop('database_type_config')
+
+    def test_type_regexp_config(self):
+        config.push('regexp_config', """\
+[mta]
+configuration: python:mailman.mta.tests.data.postfix_type_regexp
+""")
+        lmtp = LMTP()
+        self.assertEqual('regexp', lmtp.transport_file_type)
+        config.pop('regexp_config')